requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
anyio==4.7.0
    # via httpx
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
boto3==1.35.76
    # via mynotif (setup.py)
botocore==1.35.76
    # via
    #   boto3
    #   s3transfer
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
cryptography==44.0.0
    # via social-auth-core
cython==3.0.11
    # via mynotif (setup.py)
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
django==5.1.4
    # via
    #   django-cors-headers
    #   django-extensions
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
    #   mynotif (setup.py)
    #   social-auth-app-django
django-cors-headers==4.6.0
    # via mynotif (setup.py)
django-extensions==3.2.3
    # via mynotif (setup.py)
django-storages==1.14.4
    # via mynotif (setup.py)
djangorestframework==3.15.2
    # via
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   mynotif (setup.py)
djangorestframework-simplejwt==5.3.1
    # via djoser
djoser==2.3.1
    # via mynotif (setup.py)
drf-spectacular==0.28.0
    # via mynotif (setup.py)
gunicorn==23.0.0
    # via mynotif (setup.py)
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via onesignal-sdk
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-spectacular
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
onesignal-sdk==2.0.0
    # via mynotif (setup.py)
packaging==24.2
    # via gunicorn
pillow==11.0.0
    # via mynotif (setup.py)
psycopg2-binary==2.9.10
    # via mynotif (setup.py)
pycparser==2.22
    # via cffi
pyjwt==2.10.1
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   mynotif (setup.py)
python-dotenv==1.0.1
    # via mynotif (setup.py)
python3-openid==3.2.0
    # via social-auth-core
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   mynotif (setup.py)
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   requests-oauthlib
    #   social-auth-core
    #   stripe
requests-oauthlib==2.0.0
    # via social-auth-core
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.4
    # via boto3
sentry-sdk==2.19.2
    # via mynotif (setup.py)
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
social-auth-app-django==5.4.2
    # via djoser
social-auth-core==4.5.4
    # via social-auth-app-django
sqlparse==0.5.2
    # via django
stripe==11.3.0
    # via mynotif (setup.py)
typing-extensions==4.12.2
    # via stripe
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.3
    # via
    #   botocore
    #   requests
    #   sentry-sdk
whitenoise==6.8.2
    # via mynotif (setup.py)
